url_handlers: Add an /admin/stats/ to show overall receipt stats
In order to avoid code duplication of the gathering of receipt stats,
this required a modification to gather_receipt_stats_for_user() to allow
passing -1 for the uid (to signify we want to cover all users), to cater
for this the uid argument was changed to a long long, this should allow
for the -1 as well as the max unsigned int value for real uid's.

We then change our SQL queries to either include the uid clause or not.

+++ www/url_helpers.h
@@ -31,7 +31,7 @@ int do_add_user(unsigned char capabilities);
 void do_update_user(void);
 void do_edit_user(void);
 void do_activate_user(const char *uid, const char *key, const char *password);
-void gather_receipt_stats_for_user(unsigned int uid, TMPL_varlist *varlist);
+void gather_receipt_stats_for_user(long long uid, TMPL_varlist *varlist);
 void send_template(const char *template, TMPL_varlist *varlist,
 							TMPL_fmtlist *fmtlist);
 bool match_uri(const char *request_uri, const char *match);
